+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v0.12.3 - WIP
 * Get rid of Ext's viewport warning
 * Uglify components' JS when not in the test/development env
+* Fix stack-overflow issue in certain cases
 
 # v0.12.2 - 2015-06-06
 * Fix loading multi-instance components